[FEAT] item Dictionary link updata
